+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Changed
 - Update University Housing Dining section name references in favorites [#5872](https://github.com/rokwire/illinois-app/issues/5872).
+- Updated MTD Bus Stops home panel [#5880](https://github.com/rokwire/illinois-app/issues/5880).
 
 ## [8.2.4] - 2026-08-14
 ### Fixed